backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
google-generativeai
transformers
torch
python-dotenv
